--- src/components/hocs/withAuthForComponent/index.ts
@@ -1,1 +1,0 @@
-export * from "./withAuthForComponent";
